Feaure/integration pid (#66)
* WIP: Integration of PID

* WIP: Integration of PID

* Renaming

* limpieza de las funciones de hal que usaban los punteros a sensor y creación de los metodos necesarios para acceder a esas funciones.

* añadido fichero para overrides de configuracion

* fix status valve

* Reorder files and name class

* fix double

* Created new BME differencial sensor

* wip, ajuste de los PIDs

* wip - pruebas con otra libreria pid

* Add new mode and configuration to pause and resume the machine

* Fix error in innitialitation

+++ Firmware/aire_apollo/src/servoValve.h
@@ -9,13 +9,13 @@ class servoValve : public ApolloValve
 public:
   servoValve(uint8_t pin,uint8_t minpos = 0,uint8_t maxpos = 180);
   virtual bool    begin();
-  virtual void    open(uint8_t percent = 100);
+  virtual void    open(double percent = 100);
   virtual void    close();
-  virtual uint8_t status() {return _percent;}
+  virtual double status() {return _percent;}
 protected:
   Servo   _servo;
   uint8_t _pin;
-  uint8_t _percent;
+  double _percent;
   uint8_t _minPos;
   uint8_t _maxPos;
 };
